Products mostly done

+++ routes/web.php
@@ -1,7 +1,9 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\admin\ProductController;
 use App\Http\Controllers\admin\CategoryController;
+use App\Http\Controllers\admin\DashboardController;
 
 Route::get('/', function () {
     return view('welcome');
@@ -11,12 +13,21 @@ Route::get('/', function () {
 Route::get('/admin/dashboard', [App\Http\Controllers\admin\DashboardController::class, 'index'])->name('admin.dashboard');
 
 //category
-Route::get('/categories/create',[CategoryController::class,'createCategory'])->name('categories.create');
-Route::post('/categories/post', [CategoryController::class, 'storeCategory'])->name('categories.store');
-Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
-Route::delete('/categories/{id}', [CategoryController::class, 'delete'])->name('categories.delete');
-Route::get('/categories/{id}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
-Route::put('/categories/{id}', [CategoryController::class, 'update'])->name('categories.update');
+Route::get('/admin/categories/create',[CategoryController::class,'createCategory'])->name('categories.create');
+Route::post('/admin/categories/post', [CategoryController::class, 'storeCategory'])->name('categories.store');
+Route::get('/admin/categories', [CategoryController::class, 'index'])->name('categories.index');
+Route::delete('/admin/categories/{id}', [CategoryController::class, 'delete'])->name('categories.delete');
+Route::get('/admin/categories/{id}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
+Route::put('/admin/categories/{id}', [CategoryController::class, 'update'])->name('categories.update');
+
+//products
+Route::get('/admin/products/create', [ProductController::class, 'createProducts'])->name('products.create');
+Route::post('/admin/products/post', [ProductController::class, 'storeProducts'])->name('products.store');
+Route::get('/admin/products/view', [ProductController::class, 'viewProducts'])->name('products.view');
+Route::delete('/admin/products/{id}', [ProductController::class, 'deleteProducts'])->name('products.delete');
+Route::get('/admin/products/{id}/edit', [ProductController::class, 'editProducts'])->name('products.edit');
+Route::put('/admin/products/{id}', [ProductController::class, 'updateProducts'])->name('products.update');
+
 
 //order
 Route::get('/admin/order', [App\Http\Controllers\admin\OrderController::class, 'order'])->name('admin.order');
